requirements.txt
absl-py==2.2.2
addict==2.4.0
ansitable==0.11.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colored==2.3.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
dash==3.0.3
decorator==5.2.1
distlib==0.3.9
etils==1.12.2
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
fonttools==4.57.0
fsspec==2025.3.2
glfw==2.7.0
graphviz==0.20.3
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
identify==2.6.9
idna==3.10
imageio==2.36.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipython==8.35.0
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
mujoco==3.3.1
mujoco-py==2.1.2.14
narwhals==1.35.0
nbformat==5.5.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.24.4
numpy-quaternion==2023.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
open3d==0.16.0
opencv-python==4.9.0.80
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
protobuf==6.30.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
skrl==1.3.0
spatialmath-python==1.1.9
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.3.0
torchviz==0.0.3
tqdm==4.67.1
traitlets==5.14.3
triton==2.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
zipp==3.21.0
